+++ Source/Applications/Desktop/Vodovoz/MainWindow.cs
@@ -1,4 +1,4 @@
-using Autofac;
+﻿using Autofac;
 using Fias.Client;
 using Fias.Client.Cache;
 using NLog;
@@ -499,14 +499,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionSalesByDicountReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<SalesByDiscountReport>(),
-			() => new QSReport.ReportViewDlg(new SalesByDiscountReport())
-		);
-	}
-
 	protected void OnActionDriverWagesActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
@@ -526,8 +518,8 @@ public partial class MainWindow : Gtk.Window
 	protected void OnActionShortfallBattlesActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Bottles.ShortfallBattlesReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Bottles.ShortfallBattlesReport())
+			QSReport.ReportViewDlg.GenerateHashName<ShortfallBattlesReport>(),
+			() => new QSReport.ReportViewDlg(new ShortfallBattlesReport())
 		);
 	}
 
@@ -572,13 +564,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionCashierCommentsActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.Reports.CashierCommentsReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.Reports.CashierCommentsReport())
-		);
-	}
 
 	protected void OnActionCommentsForLogistsActivated(object sender, EventArgs e)
 	{
@@ -653,13 +638,7 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionOrdersByDistrict(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<OrdersByDistrictReport>(),
-			() => new QSReport.ReportViewDlg(new OrdersByDistrictReport())
-		);
-	}
+
 
 	protected void OnActionCompanyTrucksActivated(object sender, EventArgs e)
 	{
@@ -669,21 +648,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionLastOrderReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<LastOrderByDeliveryPointReport>(),
-			() => new QSReport.ReportViewDlg(new LastOrderByDeliveryPointReport())
-		);
-	}
-
-	protected void OnActionOrderIncorrectPricesReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<OrderIncorrectPrices>(),
-			() => new QSReport.ReportViewDlg(new OrderIncorrectPrices())
-		);
-	}
 
 	protected void OnActionAddressDuplicetesActivated(object sender, EventArgs e)
 	{
@@ -698,44 +662,28 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionOrdersWithMinPriceLessThanActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<OrdersWithMinPriceLessThan>(),
-			() => new QSReport.ReportViewDlg(new OrdersWithMinPriceLessThan())
-		);
-	}
 
 	protected void OnActionRouteListsOnClosingActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Logistic.RouteListsOnClosingReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Logistic.RouteListsOnClosingReport())
+			QSReport.ReportViewDlg.GenerateHashName<RouteListsOnClosingReport>(),
+			() => new QSReport.ReportViewDlg(new RouteListsOnClosingReport())
 		);
 	}
 
 	protected void OnActionOnLoadTimeActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Logistic.OnLoadTimeAtDayReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Logistic.OnLoadTimeAtDayReport())
+			QSReport.ReportViewDlg.GenerateHashName<OnLoadTimeAtDayReport>(),
+			() => new QSReport.ReportViewDlg(new OnLoadTimeAtDayReport())
 		);
 	}
-
-	protected void OnActionSelfDeliveryReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<SelfDeliveryReport>(),
-			() => new QSReport.ReportViewDlg(new SelfDeliveryReport())
-		);
-	}
-
 
 	protected void OnActionShipmentReportActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Logistic.ShipmentReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Logistic.ShipmentReport())
+			QSReport.ReportViewDlg.GenerateHashName<ShipmentReport>(),
+			() => new QSReport.ReportViewDlg(new ShipmentReport())
 		);
 	}
 
@@ -768,14 +716,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionSuburbWaterPriceActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Sales.SuburbWaterPriceReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Sales.SuburbWaterPriceReport())
-		);
-	}
-
 	protected void OnActionDistanceFromCenterActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
@@ -804,16 +744,16 @@ public partial class MainWindow : Gtk.Window
 	protected void OnActionBottlesMovementSummaryReportActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Bottles.BottlesMovementSummaryReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Bottles.BottlesMovementSummaryReport())
+			QSReport.ReportViewDlg.GenerateHashName<BottlesMovementSummaryReport>(),
+			() => new QSReport.ReportViewDlg(new BottlesMovementSummaryReport())
 		);
 	}
 
 	protected void OnActionDriveingCallsActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Logistic.DrivingCallReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Logistic.DrivingCallReport())
+			QSReport.ReportViewDlg.GenerateHashName<DrivingCallReport>(),
+			() => new QSReport.ReportViewDlg(new DrivingCallReport())
 		);
 	}
 
@@ -825,21 +765,6 @@ public partial class MainWindow : Gtk.Window
 			() => new QSReport.ReportViewDlg(new MastersVisitReport(employeeFactory))
 		);
 	}
-
-	protected void OnActionNotDeliveredOrdersActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<NotDeliveredOrdersReport>(),
-			() => new QSReport.ReportViewDlg(new NotDeliveredOrdersReport())
-		);
-	}
-
-	protected void OnActionCounterpartyTagsActivated(object sender, EventArgs e)
-	{
-		var refWin = new OrmReference(typeof(Tag));
-		tdiMain.AddTab(refWin);
-	}
-
 
 	protected void OnAction48Activated(object sender, EventArgs e)
 	{
@@ -871,27 +796,11 @@ public partial class MainWindow : Gtk.Window
 		tdiMain.AddTab(widget);
 	}
 
-	protected void OnActionOrderRegistryActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<OrderRegistryReport>(),
-			() => new QSReport.ReportViewDlg(new OrderRegistryReport())
-		);
-	}
-
 	protected void OnActionEquipmentBalanceActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Store.EquipmentBalance>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Store.EquipmentBalance())
-		);
-	}
-
-	protected void OnActionCardPaymentsActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<CardPaymentsOrdersReport>(),
-			() => new QSReport.ReportViewDlg(new CardPaymentsOrdersReport(UnitOfWorkFactory.GetDefaultFactory))
+			QSReport.ReportViewDlg.GenerateHashName<EquipmentBalance>(),
+			() => new QSReport.ReportViewDlg(new EquipmentBalance())
 		);
 	}
 
@@ -916,16 +825,6 @@ public partial class MainWindow : Gtk.Window
 		tdiMain.OpenTab(
 			PermissionControlledRepresentationJournal.GenerateHashName<TraineeVM>(),
 			() => new PermissionControlledRepresentationJournal(new TraineeVM())
-		);
-	}
-
-	protected void OnOnLineActionActivated(object sender, EventArgs e)
-	{
-		var paymentsRepository = new PaymentsRepository();
-
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<PaymentsFromTinkoffReport>(),
-			() => new QSReport.ReportViewDlg(new PaymentsFromTinkoffReport(paymentsRepository))
 		);
 	}
 
@@ -983,8 +882,8 @@ public partial class MainWindow : Gtk.Window
 	protected void OnActionOrderCreationDateReportActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<Vodovoz.ReportsParameters.Sales.OrderCreationDateReport>(),
-			() => new QSReport.ReportViewDlg(new Vodovoz.ReportsParameters.Sales.OrderCreationDateReport())
+			QSReport.ReportViewDlg.GenerateHashName<OrderCreationDateReport>(),
+			() => new QSReport.ReportViewDlg(new OrderCreationDateReport())
 		);
 	}
 
@@ -994,16 +893,6 @@ public partial class MainWindow : Gtk.Window
 			QSReport.ReportViewDlg.GenerateHashName<NotFullyLoadedRouteListsReport>(),
 			() => new QSReport.ReportViewDlg(new NotFullyLoadedRouteListsReport())
 		);
-	}
-
-	protected void OnActionFirstClientsActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<FirstClientsReport>(),
-			() => new QSReport.ReportViewDlg(
-				  new FirstClientsReport(
-						new DistrictJournalFactory(),
-						new DiscountReasonRepository())));
 	}
 
 	protected void OnActionTariffZoneDebtsReportActivated(object sender, EventArgs e)
@@ -1129,14 +1018,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnAction66Activated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<PotentialFreePromosetsReport>(),
-			() => new QSReport.ReportViewDlg(new PotentialFreePromosetsReport())
-		);
-	}
-
 	protected void OnActionWayBillReportActivated(object sender, EventArgs e)
 	{
 		tdiMain.OpenTab(
@@ -1150,15 +1031,6 @@ public partial class MainWindow : Gtk.Window
 					autofacScope.Resolve<ISubdivisionRepository>()
 				)
 			)
-		);
-	}
-
-	protected void OnActionPaymentsReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<PaymentsFromBankClientReport>(),
-			() => new QSReport.ReportViewDlg(
-				new PaymentsFromBankClientReport(new CounterpartyJournalFactory(Startup.AppDIContainer.BeginLifetimeScope()), new UserRepository(), ServicesConfig.CommonServices))
 		);
 	}
 
@@ -1243,14 +1115,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnAction71Activated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<EShopSalesReport>(),
-			() => new QSReport.ReportViewDlg(new EShopSalesReport())
-		);
-	}
-
 	protected void OnActionWayBillJournalActivated(object sender, EventArgs e)
 	{
 		var employeeFilter = new EmployeeFilterViewModel
@@ -1260,7 +1124,7 @@ public partial class MainWindow : Gtk.Window
 
 		var employeesJournalFactory = new EmployeeJournalFactory(employeeFilter);
 		var docTemplateRepository = new DocTemplateRepository();
-		var fileChooser = new Vodovoz.FileChooser();
+		var fileChooser = new FileChooser();
 
 		tdiMain.OpenTab(
 			() => new WayBillGeneratorViewModel
@@ -1386,20 +1250,6 @@ public partial class MainWindow : Gtk.Window
 		);
 	}
 
-	protected void OnActionNomenclaturePlanReportActivated(object sender, EventArgs e)
-	{
-		IProductGroupJournalFactory productGroupJournalFactory = new ProductGroupJournalFactory();
-		IParametersProvider parametersProvider = new ParametersProvider();
-		INomenclaturePlanParametersProvider nomenclaturePlanParametersProvider = new NomenclaturePlanParametersProvider(parametersProvider);
-		IFileDialogService fileDialogService = new FileDialogService();
-
-		NomenclaturePlanReportViewModel viewModel = new NomenclaturePlanReportViewModel(UnitOfWorkFactory.GetDefaultFactory,
-			ServicesConfig.InteractiveService, NavigationManager, ServicesConfig.CommonServices, productGroupJournalFactory, nomenclaturePlanParametersProvider,
-			fileDialogService);
-
-		tdiMain.AddTab(viewModel);
-	}
-
 	protected void OnLogisticsGeneralSalaryInfoActivated(object sender, EventArgs e)
 	{
 		var filter = new EmployeeFilterViewModel
@@ -1413,15 +1263,6 @@ public partial class MainWindow : Gtk.Window
 			QSReport.ReportViewDlg.GenerateHashName<GeneralSalaryInfoReport>(),
 			() => new QSReport.ReportViewDlg(new GeneralSalaryInfoReport(employeeJournalFactory, ServicesConfig.InteractiveService))
 		);
-	}
-
-	protected void OnActionOrderAnalyticsReportActivated(object sender, EventArgs e)
-	{
-		var uowFactory = autofacScope.Resolve<IUnitOfWorkFactory>();
-		var interactiveService = autofacScope.Resolve<IInteractiveService>();
-
-		NavigationManager.OpenViewModel<OrderAnalyticsReportViewModel, INavigationManager, IUnitOfWorkFactory, IInteractiveService>(
-			null, NavigationManager, uowFactory, interactiveService);
 	}
 
 	protected void OnActionEmployeesReportActivated(object sender, EventArgs e)
@@ -1499,14 +1340,6 @@ public partial class MainWindow : Gtk.Window
 		tdiMain.OpenTab(
 			QSReport.ReportViewDlg.GenerateHashName<CounterpartyCashlessDebtsReport>(),
 			() => new QSReport.ReportViewDlg(autofacScope.Resolve<CounterpartyCashlessDebtsReport>())
-		);
-	}
-
-	protected void OnActionPaymentsFromAvangardReportActivated(object sender, EventArgs e)
-	{
-		tdiMain.OpenTab(
-			QSReport.ReportViewDlg.GenerateHashName<PaymentsFromAvangardReport>(),
-			() => new QSReport.ReportViewDlg(new PaymentsFromAvangardReport())
 		);
 	}
